backend/requirements.txt
# ════════════════════════════════════════════════════════════
#  tech-fix-bible · backend requirements
#  Python >= 3.11 required
# ════════════════════════════════════════════════════════════

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database (async)
sqlalchemy==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9        # sync fallback for admin scripts if ever needed

# Auth
authlib==1.3.1                 # Google OAuth2
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0            # Starlette SessionMiddleware

# HTTP client for link health check
httpx==0.27.0

# Scheduling
apscheduler==3.10.4

# Config / utilities
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4

# Dev (optional — uncomment for local)
# pytest==8.2.2
# pytest-asyncio==0.23.7
# ruff==0.5.0
